requirements.txt
astroid==2.1.0
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
Django==2.1.5
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
idna==2.7
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.2
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.2.2
python-dotenv==0.10.1
pytz==2018.9
requests==2.20.1
six==1.12.0
traitlets==4.3.2
typed-ast==1.1.1
urllib3==1.24.1
wcwidth==0.1.7
wrapt==1.11.0
django-extensions==2.1.5